% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{as_nc}
\alias{as_nc}
\title{Revert to NetCDF connection class}
\usage{
as_nc(x, ...)
}
\arguments{
\item{x}{open_dap connection}

\item{...}{ignored}
}
\value{
NetCDF connection
}
\description{
Revert to NetCDF connection class
}
