% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{canon_dap_url}
\alias{canon_dap_url}
\title{OpenDAP example URL}
\usage{
canon_dap_url(...)
}
\arguments{
\item{...}{ignored}
}
\value{
character string, returns a test URL that should work if all is well
}
\description{
OpenDAP example URL
}
\examples{
canon_dap_url()
}
