% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{open_dap}
\alias{open_dap}
\alias{open_dap}
\alias{print.open_dap}
\alias{open_dap}
\alias{dap_live}
\title{Open a connection}
\usage{
open_dap(src, ...)

\method{print}{open_dap}(x, ...)

dap_live(x, ...)
}
\arguments{
\item{src}{OpenDAP source string}

\item{...}{arguments passed to `RNetCDF::open.nc`, `share` and `prefill`}
}
\value{
open_dap connection, a NetCDF connection with some extras
}
\description{
OpenDAP connection
}
